[FIX] creature: enable lvl6_up creature frenzy ability
Issue:
Casting the original `spell.frenzy` as the lvl6_up creature ability does not
seem to work.

Fix:
Added a new `frenzyAbility` spell which will mimick the behavior of
basic frenzy.

Known issues after fix:
Interface issue, the creature bonuses display shows that the frenzy bonus has
a 3 turns duration, but this is not true, the bonus is properly removed at
the start of the next action turn of the creature stack.